Feature download scripts

+++ pyspectral/utils.py
@@ -336,7 +336,7 @@ def convert2hdf5(ClassIn, platform_name, bandnames, scale=1e-06):
             dset[...] = arr
 
 
-def download_rsr():
+def download_rsr(**kwargs):
     """Download the pre-compiled hdf5 formatet relative spectral response functions
     from the internet
 
@@ -345,16 +345,35 @@ def download_rsr():
     #
     import tarfile
     import requests
-    from tqdm import tqdm
+    TQDM_LOADED = True
+    try:
+        from tqdm import tqdm
+    except ImportError:
+        TQDM_LOADED = False
+
+    dest_dir = kwargs.get('dest_dir', LOCAL_RSR_DIR)
+    dry_run = kwargs.get('dry_run', False)
+
+    LOG.info("Download RSR files and store in directory %s", dest_dir)
+
+    filename = os.path.join(dest_dir, "pyspectral_rsr_data.tgz")
+    LOG.debug("Get data. URL: %s", HTTP_PYSPECTRAL_RSR)
+    LOG.debug("Destination = %s", dest_dir)
+    if dry_run:
+        return
 
     response = requests.get(HTTP_PYSPECTRAL_RSR)
-    filename = os.path.join(LOCAL_RSR_DIR, "pyspectral_rsr_data.tgz")
-    with open(filename, "wb") as handle:
-        for data in tqdm(response.iter_content()):
-            handle.write(data)
+    if TQDM_LOADED:
+        with open(filename, "wb") as handle:
+            for data in tqdm(response.iter_content()):
+                handle.write(data)
+    else:
+        with open(filename, "wb") as handle:
+            for data in response.iter_content():
+                handle.write(data)
 
     tar = tarfile.open(filename)
-    tar.extractall(LOCAL_RSR_DIR)
+    tar.extractall(dest_dir)
     tar.close()
     os.remove(filename)
 
@@ -370,6 +389,8 @@ def download_luts(**kwargs):
     except ImportError:
         TQDM_LOADED = False
 
+    dry_run = kwargs.get('dry_run', False)
+
     if 'aerosol_type' in kwargs:
         if isinstance(kwargs['aerosol_type'], (list, tuple, set)):
             aerosol_types = kwargs['aerosol_type']
@@ -382,20 +403,27 @@ def download_luts(**kwargs):
 
     for subname in aerosol_types:
 
+        LOG.debug('Aerosol type: %s', subname)
         http = HTTPS_RAYLEIGH_LUTS[subname]
+        LOG.debug('URL = %s', http)
 
+        subdir_path = RAYLEIGH_LUT_DIRS[subname]
         try:
-            os.makedirs(RAYLEIGH_LUT_DIRS[subname])
+            LOG.debug('Create directory: %s', subdir_path)
+            if not dry_run:
+                os.makedirs(subdir_path)
         except OSError:
-            if not os.path.isdir(RAYLEIGH_LUT_DIRS[subname]):
+            if not os.path.isdir(subdir_path):
                 raise
+
+        if dry_run:
+            continue
 
         response = requests.get(http)
         total_size = int(response.headers['content-length'])
 
-        subdirname = RAYLEIGH_LUT_DIRS[subname]
         filename = os.path.join(
-            subdirname, "pyspectral_rayleigh_correction_luts.tgz")
+            subdir_path, "pyspectral_rayleigh_correction_luts.tgz")
         if TQDM_LOADED:
             with open(filename, "wb") as handle:
                 for data in tqdm(iterable=response.iter_content(chunk_size=chunk_size),
@@ -407,7 +435,7 @@ def download_luts(**kwargs):
                     handle.write(data)
 
         tar = tarfile.open(filename)
-        tar.extractall(subdirname)
+        tar.extractall(subdir_path)
         tar.close()
         os.remove(filename)
